Change controller tests to mock out interactor usage
The tests were already growing in such a way as to require knowledge
of how the entire application worked. I don't want to have to deal with
that in tests. The Rails side should assume the app works, and thus
I am now mocking out interactor usage in controller tests