Replace scoped 'within' steps with straight element checks
I've never been a fan of the selectors.rb idea in websteps. It's to easy to
forget what's there and end up duplicating logic, or creating a massive set
of selectors to names that becomes difficult to maintain. In terms of seeing
how low level cucumber tests work I'm going to use the raw CSS selectors
instead.