Pull Raid form into it's own partial and use form_for
Because of the change to form_for, we also need to send a Raid object
down into the form. As a side effect of the #edit work (next) also
including object values in the form.

I'm not using form_for helpers right now as I don't want to hook myself
strongly to the fields of my domain models. This will probably change.