Further refactor ChangeUser to not cause double-save
Realizing that I'm saving the user twice here, which due to the objects
being the same one of the saves is redundant. Also, this new object is
helping build up a vocabulary, where Update[x] objects actually do the
saving to persistence and Change[x] just update the object without
saving.