Refactor in-memory to dup objects when loading and saving
Finally got around to an Update that had error handling and found out a
small issue with my in-memory impl: there was only one object for a
given record. This meant that any change made to the object was changed
everywhere immediately, and Repository#save was never really required.

The refactor changes the internal impl to require save and load, and
properly dups the objects when saving *and* loading to ensure that at
any point we are working with the object we asked for and aren't going
to accidentally change the stored version.

There are also a number of test failures with this change that I'll be
fixing up next.