Refactor AddCharacter to take an attributes hash
For actions that deal with direct creation or update of a model, the
Rails-way is feeling cleaner and easier to control and maintain. Whether
this is just "hiding the problem" will remain to be seen but for now I'm
working to keep the number of parameters to a minimum so they're easily
documented.