Refactor guild->[char] mapping to it's own top-level object
After fighting with trying to get the guild -> [characters] mapping
working in it's current context I realized that not only do I need this
functionality elsewhere but that I'm forcing too much on ListCharacters
itself. I've pulled the inner class out to it's own top-level location,
got it tested, and hooked it up to the Characters page as well.

The "remove characters already signed up" logic is still in the
controller, which I'm not the happiest about but this is much cleaner
than it was before, and reusable, so I'm happy.